+++ changelog.markdown
@@ -1,5 +1,6 @@
 ## 0.4.0
  * the match option on filters takes symbols
+ * adding support for hidden filters
 
 ## 0.3.2
  * fix join table name finder in find_column_type